requirements.txt
ansicon==1.89.0
arabic-reshaper==3.0.0
arrow==1.2.3
asgiref @ file:///C:/Users/BUILDE~1/AppData/Local/Temp/abs_cchykh5sgr/croots/recipe/asgiref_1661287542702/work
asn1crypto==1.5.1
astroid==2.15.0
async-timeout==4.0.2
autopep8==2.0.2
black==23.1.0
blessed==1.20.0
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
crispy-bootstrap4==2022.1
cryptography==39.0.2
cssselect2==0.7.0
defusedxml==0.7.1
dill==0.3.6
Django @ file:///tmp/build/80754af9/django_1617740316746/work
django-allauth==0.52.0
django-appconf==1.0.5
django-auto-logout==0.5.1
django-autoslug==1.9.8
django-ckeditor==6.5.1
django-compressor==4.3.1
django-crispy-forms==2.0
django-environ==0.10.0
django-extensions==3.2.1
django-filebrowser-no-grappelli==4.0.1
django-fontawesome==1.0
django-js-asset==2.0.0
django-pandas==0.6.6
django-phonenumber-field==7.0.2
django-phonenumbers==1.0.1
django-picklefield==3.1
django-q==1.3.9
django-redis==5.2.0
django-silk==5.0.3
django-simple-history==3.2.0
flake8==6.0.0
gprof2dot==2022.7.29
html5lib==1.1
idna==3.4
isort==5.12.0
jinxed==1.2.0
lazy-object-proxy==1.9.0
lxml==4.9.2
mccabe==0.7.0
mypy-extensions==1.0.0
numpy==1.24.2
oauthlib==3.2.2
oscrypto==1.3.0
packaging==23.0
pandas==1.5.3
pathspec==0.11.0
phonenumbers==8.13.7
Pillow==9.4.0
platformdirs==3.1.0
psycopg-binary==3.1.8
psycopg2 @ file:///C:/b/abs_6cdgqr8yu7/croot/psycopg2_1666817457048/work
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
pyHanko==0.17.0
pyhanko-certvalidator==0.20.1
PyJWT==2.6.0
pylint==2.16.4
pypdf==3.5.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz @ file:///C:/b/abs_22fofvpn1x/croot/pytz_1671698059864/work
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.4.2
rcssmin==1.1.1
redis==3.5.3
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
rjsmin==1.2.1
six==1.16.0
sqlparse @ file:///C:/b/abs_83jn6owfpc/croot/sqlparse_1668793499314/work
svglib==1.5.1
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uritools==4.0.1
urllib3==1.26.14
wcwidth==0.2.6
webencodings==0.5.1
wincertstore==0.2
wrapt==1.15.0
xhtml2pdf==0.2.9
